+++ tests/messages_data/rfc2822/example03.py
@@ -14,9 +14,9 @@ DATA = dict(
     html='',
     headers={'from': ('"Joe Q. Public" <john.q.public@example.com>',), 'to': ('Mary Smith <mary@x.test>, jdoe@example.org, Who? <one@y.test>',), 'cc': ('<boss@nil.test>, "Giant; \\"Big\\" Box" <sysservices@example.net>',), 'date': ('Tue, 1 Jul 2003 10:52:37 +0200',), 'message-id': ('<5678.21-Nov-1997@example.com>',)},
     attachments=[],
-    from_values=EmailAddress('Joe Q. Public', 'john.q.public@example.com', 'Joe Q. Public <john.q.public@example.com>'),
-    to_values=(EmailAddress('Mary Smith', 'mary@x.test', 'Mary Smith <mary@x.test>'), EmailAddress('', 'jdoe@example.org', 'jdoe@example.org'), EmailAddress('Who?', 'one@y.test', 'Who? <one@y.test>')),
-    cc_values=(EmailAddress('', 'boss@nil.test', 'boss@nil.test'), EmailAddress('Giant; "Big" Box', 'sysservices@example.net', 'Giant; "Big" Box <sysservices@example.net>')),
+    from_values=EmailAddress(name='Joe Q. Public', email='john.q.public@example.com'),
+    to_values=(EmailAddress(name='Mary Smith', email='mary@x.test'), EmailAddress(name='', email='jdoe@example.org'), EmailAddress(name='Who?', email='one@y.test')),
+    cc_values=(EmailAddress(name='', email='boss@nil.test'), EmailAddress(name='Giant; "Big" Box', email='sysservices@example.net')),
     bcc_values=(),
     reply_to_values=(),
 )